+++ opcodes.hh
@@ -141,7 +141,7 @@ enum class Opcode {
 struct Instruction {
     Opcode opcode;
     uint32_t line;
-    int32_t intValue = 0;
+    int64_t intValue = 0;
     float floatValue = 0.0f;
     bool boolValue = false;
     std::string stringValue;
